refactor(jazz-tools): consume jazz-tools/shared in the use-all bindings
Repoint the React, Svelte and Vue useAll bindings from deep relative
imports to the public jazz-tools/shared entry point. Pure import swap —
the barrel re-exports the identical symbols, so behaviour is unchanged.
Dogfooding the entry point is the proof its surface is complete.
